+++ pea2pea/src/node.rs
@@ -112,7 +112,7 @@ impl Node {
     pub fn new(mut config: Config) -> Self {
         // if there is no pre-configured name, assign a sequential numeric identifier
         if config.name.is_none() {
-            config.name = Some(SEQUENTIAL_NODE_ID.fetch_add(1, SeqCst).to_string());
+            config.name = Some(SEQUENTIAL_NODE_ID.fetch_add(1, Relaxed).to_string());
         }
 
         // create a tracing span containing the node's name